requirements.txt
flask==2.3.3
requests==2.31.0
google-genai==1.11.0
gunicorn==21.2.0 